requirements.txt
Django==2.1.3
django-render-block==0.5
django-templated-email==2.3.0
Pillow==5.3.0
pytz==2018.7
six==1.12.0
